requirements.txt
flask==3.0.3
datetime==5.5
PyJWT==2.8.0
cryptography==42.0.5
requests==2.31.0
beautifulsoup4~=4.12.3
selenium~=4.19.0